no more getting adamantium
on maps with area less than 1000, if an HQ sees a mana well they will make all 4 miners get the mana well. On a map with size >= 1000, if an HQ sees a mana well, they will go 3-1 with mana/adamantium. If an HQ does not see a mana well, it will always go 2-2, prioritizing adamantium first